+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- feat(uploads): support streaming uploads with on-demand part URLs
 
 ## [0.5.0] - 2026-06-26